Update 2015/Day24/day24.c
The function signature for mul64 takes an unsigned long b, but the implementation logic is only correct if b is a 16-bit value (an unsigned int on this platform). Multiplying two 16-bit values results in a 32-bit value, which fits in an unsigned long. However, if b were a true 32-bit unsigned long, a0 * b would produce up to a 48-bit result, causing an overflow in r0. Since this function is only called with unsigned int weights, you should change the signature to void mul64(uint64_t_c64* a, unsigned int b) to enforce this constraint and prevent future bugs.

+++ 2015/Day24/day24.c
@@ -30,7 +30,7 @@ void copy64(uint64_t_c64* dest, uint64_t_c64* src) {
 }
 
 /* 64-bit multiplication: a = a * b */
-void mul64(uint64_t_c64* a, unsigned long b) {
+void mul64(uint64_t_c64* a, unsigned int b) {
     unsigned long r0, r1, r2, r3;
     unsigned long a0 = a->low & 0xFFFF;
     unsigned long a1 = (a->low >> 16) & 0xFFFF;
